Add the render function to render ReST
recliner.render can be used to take source text and render it to
html. It uses a standard set of settings that is both featurefull
(syntax highlight, Math) and safe (disables raw html, file insertion).

It also normalizes exceptions to ValueError, quiets docutils from
yelling in sys.stderr, and makes any errors in markup result in
an exception instead of missing segments.